+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/entity-detail/entity-detail-workspace-base.ts
@@ -1,5 +1,6 @@
 import { UmbSubmittableWorkspaceContextBase } from '../submittable/index.js';
 import { UmbEntityWorkspaceDataManager } from '../entity/entity-workspace-data-manager.js';
+import type { UmbEntityDetailWorkspaceContextArgs, UmbEntityDetailWorkspaceContextCreateArgs } from './types.js';
 import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbEntityContext, type UmbEntityModel, type UmbEntityUnique } from '@umbraco-cms/backoffice/entity';
@@ -14,21 +15,7 @@ import { umbExtensionsRegistry, type ManifestRepository } from '@umbraco-cms/bac
 import type { UmbDetailRepository } from '@umbraco-cms/backoffice/repository';
 import { UmbStateManager } from '@umbraco-cms/backoffice/utils';
 
-export interface UmbEntityDetailWorkspaceContextArgs {
-	entityType: string;
-	workspaceAlias: string;
-	detailRepositoryAlias: string;
-}
-
-/**
- * @deprecated Use UmbEntityDetailWorkspaceContextArgs instead
- */
-export type UmbEntityWorkspaceContextArgs = UmbEntityDetailWorkspaceContextArgs;
-
-export interface UmbEntityDetailWorkspaceContextCreateArgs<DetailModelType> {
-	parent: UmbEntityModel;
-	preset?: Partial<DetailModelType>;
-}
+const LOADING_STATE_UNIQUE = 'umbLoadingEntityDetail';
 
 export abstract class UmbEntityDetailWorkspaceContextBase<
 	DetailModelType extends UmbEntityModel = UmbEntityModel,
@@ -61,8 +48,6 @@ export abstract class UmbEntityDetailWorkspaceContextBase<
 	public readonly parentEntityType = this.#parent.asObservablePart((parent) =>
 		parent ? parent.entityType : undefined,
 	);
-
-	#loadingStateUnique = 'umbLoadingEntityDetail';
 
 	#initResolver?: () => void;
 	#initialized = false;
@@ -138,7 +123,7 @@ export abstract class UmbEntityDetailWorkspaceContextBase<
 
 	async load(unique: string) {
 		this.#entityContext.setUnique(unique);
-		this.loading.addState({ unique: this.#loadingStateUnique, message: `Loading ${this.getEntityType()} Details` });
+		this.loading.addState({ unique: LOADING_STATE_UNIQUE, message: `Loading ${this.getEntityType()} Details` });
 		await this.#init;
 		this.resetState();
 		this._getDataPromise = this._detailRepository!.requestByUnique(unique);
@@ -158,7 +143,7 @@ export abstract class UmbEntityDetailWorkspaceContextBase<
 			);
 		}
 
-		this.loading.removeState(this.#loadingStateUnique);
+		this.loading.removeState(LOADING_STATE_UNIQUE);
 		return response;
 	}
 
@@ -181,7 +166,7 @@ export abstract class UmbEntityDetailWorkspaceContextBase<
 	 * @returns { Promise<any> | undefined } The data of the scaffold.
 	 */
 	public async createScaffold(args: CreateArgsType) {
-		this.loading.addState({ unique: this.#loadingStateUnique, message: `Creating ${this.getEntityType()} scaffold` });
+		this.loading.addState({ unique: LOADING_STATE_UNIQUE, message: `Creating ${this.getEntityType()} scaffold` });
 		await this.#init;
 		this.resetState();
 		this.setParent(args.parent);
@@ -202,7 +187,7 @@ export abstract class UmbEntityDetailWorkspaceContextBase<
 			this._data.setCurrent(data);
 		}
 
-		this.loading.removeState(this.#loadingStateUnique);
+		this.loading.removeState(LOADING_STATE_UNIQUE);
 
 		return data;
 	}
@@ -303,9 +288,9 @@ export abstract class UmbEntityDetailWorkspaceContextBase<
 		}
 
 		if (this._checkWillNavigateAway(newUrl) && this._getHasUnpersistedChanges()) {
-			/* Since ours modals are async while events are synchronous, we need to prevent the default behavior of the event, even if the modal hasn’t been resolved yet. 
-			Once the modal is resolved (the user accepted to discard the changes and navigate away from the route), we will push a new history state. 
-			This push will make the "willchangestate" event happen again and due to this somewhat "backward" behavior, 
+			/* Since ours modals are async while events are synchronous, we need to prevent the default behavior of the event, even if the modal hasn’t been resolved yet.
+			Once the modal is resolved (the user accepted to discard the changes and navigate away from the route), we will push a new history state.
+			This push will make the "willchangestate" event happen again and due to this somewhat "backward" behavior,
 			we set an "allowNavigateAway"-flag to prevent the "discard-changes" functionality from running in a loop.*/
 			e.preventDefault();
 			const modalManager = await this.getContext(UMB_MODAL_MANAGER_CONTEXT);
