+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/entity-detail/index.ts
@@ -2,3 +2,4 @@ import './global-components/index.js';
 
 export * from './entity-detail-workspace-base.js';
 export * from './global-components/index.js';
+export type * from './types.js';
