+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/entity-detail/types.ts
@@ -1,0 +1,17 @@
+import type { UmbEntityModel } from '@umbraco-cms/backoffice/entity';
+
+export interface UmbEntityDetailWorkspaceContextArgs {
+	entityType: string;
+	workspaceAlias: string;
+	detailRepositoryAlias: string;
+}
+
+/**
+ * @deprecated Use UmbEntityDetailWorkspaceContextArgs instead
+ */
+export type UmbEntityWorkspaceContextArgs = UmbEntityDetailWorkspaceContextArgs;
+
+export interface UmbEntityDetailWorkspaceContextCreateArgs<DetailModelType> {
+	parent: UmbEntityModel;
+	preset?: Partial<DetailModelType>;
+}
